refactor: rename survey editor logic
